+++ management/commands/inbox_worker.py
@@ -23,83 +23,20 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import logging
-import traceback
-import uuid
-from decimal import Decimal
 from time import sleep
 
-import cattr
-from django.conf import settings
 from django.core.management import BaseCommand
-from django.db import transaction
 
-from infrastructure.messages_bus import message_bus_instance
-from osis_common.utils.thread import ConsumerThreadWorkerStrategy, OneThreadPerBoundedContextRunner
-
-logger = logging.getLogger(settings.DEFAULT_LOGGER)
-
-
-cattr.register_structure_hook(uuid.UUID, lambda d, t: d)
-cattr.register_structure_hook(Decimal, lambda d, t: d)
-
-
-class EventClassNotFound(Exception):
-    def __init__(self, event_name: str, **kwargs):
-        self.message = f"Cannot process {event_name} events because not found in handlers definition..."
-        super().__init__(self.message)
+from osis_common.utils.inbox_outbox import ConsumerThreadWorkerStrategy, OneThreadPerBoundedContextRunner, \
+    InboxConsumer
 
 
 class InboxThreadWorker(ConsumerThreadWorkerStrategy):
 
     def run(self):
         while True:
-            self._execute_unprocessed_events()
+            InboxConsumer(self.bounded_context_name, self.event_handlers).consume_all_unprocessed_events()
             sleep(5)
-
-    def _execute_unprocessed_events(self):
-        with transaction.atomic():
-            unprocessed_events = self.inbox_model.objects.select_for_update().filter(
-                consumer=self.bounded_context_name,
-                status=self.inbox_model.PENDING
-            ).order_by('creation_date')
-            if unprocessed_events:
-                logger.info(f"{self._logger_prefix_message()}: Process {len(unprocessed_events)} events...")
-
-            for unprocessed_event in unprocessed_events:
-                event_name = unprocessed_event.event_name
-                try:
-                    event_instance = self._build_event_instance(unprocessed_event)
-                    for function in self.event_handlers[event_instance.__class__]:
-                        function(message_bus_instance, event_instance)
-                    unprocessed_event.mark_as_processed()
-                except EventClassNotFound as e:
-                    logger.warning(e.message)
-                    unprocessed_event.mark_as_error(e.message)
-                    continue
-                except Exception:
-                    logger.error(
-                        f"{self._logger_prefix_message()}: Cannot process {event_name} event ({event_instance})",
-                        exc_info=True
-                    )
-                    unprocessed_event.mark_as_error(traceback.format_exc())
-                    continue
-
-    def _build_event_instance(self, unprocessed_event):
-        try:
-            event_cls = next(
-                event_cls for event_cls, fn_list in self.event_handlers.items()
-                if event_cls.__name__ == unprocessed_event.event_name
-            )
-            return cattr.structure({
-                'transaction_id': unprocessed_event.transaction_id,
-                **unprocessed_event.payload,
-            }, event_cls)
-        except StopIteration:
-            raise EventClassNotFound(unprocessed_event.event_name)
-
-    def _logger_prefix_message(self) -> str:
-        return f"[Inbox Worker - {self.bounded_context_name} - Thread: {self.ident}]"
 
 
 class Command(BaseCommand):
